requirements.txt
blinker==1.9.0
click==8.2.1
colorlog==6.9.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
psutil==7.0.0
pytz==2025.2
PyYAML==6.0.2
RPi.bme280==0.2.4
smbus2==0.5.0
Werkzeug==3.1.3
